+++ CHANGELOG.md
@@ -33,11 +33,13 @@ Resource data will be available with the resource object. This enhancement helps
 - Interconnect type
 - Internal link set
 - Logical enclosure
+- Logical interconnect group
 - Managed SAN
 - SAS interconnect
 - SAS interconnect type
 - SAS logical interconnect
 - SAS logical interconnect group
+- Server hardware
 - Server hardware type
 - Server profile template
 - Switch type